+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ================
 * Your contribution here.
 * [#527](https://github.com/oshi/oshi/pull/527): Correct process information caching and command line retrieval under Windows - [@dustin-johnson](https://github.com/dustin-johnson).
+* [#533](https://github.com/oshi/oshi/pull/533): Filter to CPU zone if multiple Windows Thermal sensors. - [@dbwiddis](https://github.com/dbwiddis).
 
 3.6.0 (6/20/2018)
 ================